// source/phone_otp_version/nestjs/src/core/utils/enums.ts
export enum MailType {
    VerifyEmail = "verifyEmail",
    ResetPassword = "resetPassword"
}

export enum VPushProvider {
    fcm = "fcm", onesignal = "onesignal"
}
export enum ChatRequestStatus {
    Accepted = "accepted",
    Refused = "refused",
    Canceled = "canceled",
    Pending = "pending"
}
export enum UserPrivacyTypes{
    ForReq="forReq",
    Public="public",
}
export enum UserRole {
    Prime = "prime",
    Admin = "admin",
    HasBadge = "hasBadge",
}

export enum AccessTokenType {
    Access = "access",
    Refresh = "refresh",
}


export enum NotificationType {
    Admin = "admin"

}


export enum TokenType {
    Access = "access",
    Refresh = "refresh"
}

export enum S3UploaderTypes {
    profileImage = "pic100",
    chatImage = "ci500",
    chatVideo = "cv501",
    chatFile = "cf502",
    chatVoice = "cv503",
    smallImage = "si400",
    mediumImage = "mi500",
    media = "media600",
}


export enum Platform {
    Android = "android",
    Ios = "ios",
    Web = "web",
    Linux = "linux",
    Mac = "macOs",
    Windows = "windows",
    Other = "other"
}

export enum PushTopics {
    AdminAndroid = "admin-android",
    AdminIos = "ios-android",
    Verified = "verified",
    UnVerified = "unverified",
}


export enum RegisterMethod {
    email = "email",
    phone = "phone",
    facebook = "facebook",
    gmail = "gmail",
    apple = "apple",
    qrCode = "qrCode",
}

export enum RegisterStatus {
    accepted = "accepted",
    pending = "pending",
    notAccepted = "notAccepted",
}